+++ src/components/drawer/DrawerItems.tsx
@@ -3,7 +3,7 @@ import {DataSourceFilter} from "./DataSourceFilter.tsx";
 import {AdvancedFilters} from "./AdvancedFilters.tsx";
 import {RadarOptions} from "./RadarOptions.tsx";
 import {SelectedFlightData} from "./SelectedFlightData.tsx";
-import {GraphicOptions} from "./GraphicOptions.tsx";
+import {GraphicOptionsDrawer} from "./GraphicOptionsDrawer.tsx";
 import {Miscellaneous} from "./Miscellaneous.tsx";
 import {SunAstronomy} from "./SunAstronomy.tsx";
 import {MoonAstronomy} from "./MoonAstronomy.tsx";
@@ -13,7 +13,7 @@ export function DrawerItems() {
         <List>
             <DataSourceFilter/>
             <RadarOptions/>
-            <GraphicOptions/>
+            <GraphicOptionsDrawer/>
             <AdvancedFilters/>
             <SelectedFlightData/>
             <Miscellaneous/>